feat: rename style files in login and register
