+++ client/src/pages/authentication/register/styles.ts
